fix: Complement ImageModel and related events
The `ImageModel` class now includes a new property `UploaderId`,
which is set to `e.Uploader.Value` in the `Apply` method.

In the `Album` class, the `ImageAddedEvent` constructor call
was modified to replace `command.Actor.Id` with `_author` and
to add `command.Actor.Id` as a new parameter.

The `ImageAddedEvent` record class was updated to include a
new property `Uploader` of type `UserId`, capturing the ID of
the user who uploaded the image.